LayoutTests/imported/w3c/web-platform-tests/streams/piping/pipe-through.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL pipe-through.js shared worker wrapper file ReferenceError: Can't find variable: SharedWorker
